+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 **Fixed:**
 * Aerial dodge infinite jump.
+* Player being able to move while in magic cast time.
 * Command Menu Text offset being too close to the left by default.
 * Material command not working when specifying a material.
 * Command Menu Magic and Drive back to ??? if none have been learned yet.